requirements.txt
Django>=3.0
psycopg2-binary>=2.8
djangorestframework
celery
django-celery-beat
feedparser
pyyaml
uritemplate
drf-spectacular